Fix merge conflicts with entities package rename
