+++ custom_components/catlink/binary_sensor.py
@@ -9,7 +9,7 @@ from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
 from .const import DOMAIN
-from .entitites import CatlinkBinaryEntity
+from .entities import CatlinkBinaryEntity
 from .helpers import Helper
