+++ custom_components/catlink/entities/catlink_binary.py
