+++ custom_components/catlink/sensor.py
@@ -9,7 +9,7 @@ from homeassistant.helpers import config_validation as cv, entity_platform
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
 from .const import DOMAIN
-from .entitites import CatlinkEntity
+from .entities import CatlinkEntity
 from .helpers import Helper
 
 async_setup_accounts = Helper.async_setup_accounts
