+++ custom_components/catlink/switch.py
@@ -8,7 +8,7 @@ from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
 from .const import DOMAIN
-from .entitites import CatlinkBinaryEntity
+from .entities import CatlinkBinaryEntity
 from .helpers import Helper
